% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/band-label.R
\name{summarize_imagery}
\alias{summarize_imagery}
\title{Turn a imagery into a rectagular file}
\usage{
summarize_imagery(path = "data", ext = ".TIF")
}
\arguments{
\item{path}{to directory}

\item{ext}{extension to select on}
}
\description{
Turn a imagery into a rectagular file
}
